Improved asset system. Keyboard events. Tentative save formats.
Assets are now loaded by the engine before the game starts, not
when the header is loaded. These are practically the same time
when you're playing the game, but it means Pygame doesn't have
to be fully set up to modify the level Clusters. Also, sounds can
be loaded.

The events system has been extended a bit to add window-level
handlers and key up/down events.

There is a system to create a SaveFile containing a Cluster and a
bit of metadata extracted from it, and to load the file afterwards.
There is also a more basic save system which creates an alphanumeric
case sensitive string containing the player's chosen name and the
level filename, both obfuscated of course.

There's a couple of other things, like HistLib copies productName
in case you want to use it in another file.