Less code duplication, simpler assets, saneobject base class, fonts
* CString and OString now both derive from a base class called
BString which contains their common code. Since the differences
between them basically amounted to boilerplate I moved their
declarations into an exec loop.
* The asset loader doesn't store the filenames of assets any more.
This means that the "Asset" struct was also removed. While I was
at it, I got rid of the loadGraphic and loadSound subroutines as
they were both one-liners. Each loader subroutine is now a member
of the "loaders" dict, making it much easier to add new ones.
* HistLib now has a nice new type called saneobject which has all
of the benefits of "new-style" classes but does not screw up the
string representation of the class object itself.
* I added a "Font" class which wraps Pygame's one. Mine is basically
a dynamic dictionary containing every possible size of a typeface
as Pygame objects, but created on access (or, if necessary, in
advance using initSize()), so, for example:

tahoma = Fonts.Font("tahoma.otf")
tahoma8 = tahoma[8]
tahoma14 = tahoma[14]

has the same result as:

tahoma8 = pygame.font.Font("tahoma.otf", 8)
tahoma14 = pygame.font.Font("tahoma.otf", 14)

Behind-the-scenes, the font file itself is loaded on creation and
then fed to Pygame using cStringIO to generate the different sizes,
or, in other words, it's only loaded once. The class has been added
as an asset loader.