Rudimentary RTF Parser
Eventually I want to have text rendering advanced enough that button
labels and such can be given in RTF (perhaps DolDoc and HTML will
also be supported). This RTF parser ignores most properties of the
data it receives and can't save, but I will hopefully be able to
improve it with time.