Bad Git Practices
Main event is fonts. Most of the text in the launcher is rendered
at startup. The new objects in the Launcher are unresponsive
because my current event system is a real stinker. I need to find
a better way to handle events, but it's giving me a headache.

There are a bunch of other minor changes. Read the diff yourself.